+++ modules/ui/sections/preset_fields.js
@@ -4,7 +4,6 @@ import { presetManager } from '../../presets';
 import { t, localizer } from '../../core/localizer';
 import { utilArrayIdentical } from '../../util/array';
 import { utilArrayUnion, utilRebind } from '../../util';
-import { modeBrowse } from '../../modes/browse';
 import { uiField } from '../field';
 import { uiFormFields } from '../form_fields';
 import { uiSection } from '../section';
@@ -117,17 +116,6 @@ export function uiSectionPresetFields(context) {
                 .state(_state)
                 .klass('grouped-items-area')
             );
-
-
-        selection.selectAll('.wrap-form-field input')
-            .on('keydown', function(d3_event) {
-                // if user presses enter, and combobox is not active, accept edits..
-                if (d3_event.keyCode === 13 && // ↩ Return
-                    context.container().select('.combobox').empty()) {
-
-                    context.enter(modeBrowse(context));
-                }
-            });
     }
 
     section.presets = function(val) {